requirements.txt
matplotlib
nltk
numpy
pandas
plotly
PyYAML
PyYAML
scikit_learn
scipy
seaborn
streamlit
